tidied up MarkdownConfluenceSync (but not yet finished)
