+++ bin/PageInfo.py
@@ -1,0 +1,8 @@
+'''
+Created on Feb 2, 2017
+
+@author: tobias-vogel-seerene
+'''
+import collections
+
+PageInfo = collections.namedtuple('PageInfo', ['id', 'version', 'link'])
